requirements.txt
Mastodon.py>=1.8.1,<2.0
python-dotenv>=1.0.1,<2.0
